Moved control_ren to a mixin and added it to USB, GPIB and TCPIP
